+++ test/serverTest.js
@@ -267,4 +267,214 @@ describe('Brave Sensor server', () => {
 
     })
 
+    describe('Smart Things API endpoint', () => {
+        it('should return 200 for a valid request', async ()=> {
+            // example valid request obtained from Smart Things documentation
+            const goodRequest = {
+                lifecycle: "PING",
+                executionId: "b328f242-c602-4204-8d73-33c48ae180af",
+                locale: "en",
+                version: "1.0.0",
+                pingData: {
+                    challenge: "1a904d57-4fab-4b15-a11e-1c4bfe7cb502"
+                }
+            }
+            
+            const response = await chai.request(server).post('/api/st').send(goodRequest)
+            expect(response).to.have.status(200)
+        })
+
+        it('should return 400 for a request that does not contain lifecycle, ie. which does not pass express validator', async () => {
+            const badExpressRequest = {
+                uselessField: 'useless'
+            }
+            
+            const response = await chai.request(server).post('/api/st').send(badExpressRequest)
+            expect(response).to.have.status(400)
+        })
+        
+        it(`should return 401 for a request that contains lifecycle (passes express validator) but which is not valid for Smart Things' purposes`, async () => {
+            const badSmartThingsRequest = {
+                lifecycle: "thisisnotvalid",
+                executionId: "b328f242-c602-4204-8d73-33c48ae180af",
+                locale: "en",
+                version: "1.0.0",
+                pingData: {
+                    challenge: "1a904d57-4fab-4b15-a11e-1c4bfe7cb502"
+                }
+            }
+            
+            const response = await chai.request(server).post('/api/st').send(badSmartThingsRequest)
+            expect(response).to.have.status(401)
+        })
+    })
+
+    describe('Express validation of API endpoints', () => {
+        describe('api/doorTest endpoint', () => {
+            beforeEach(async function() {
+                await redis.clearKeys()
+                await db.clearSessions()
+                await db.clearLocations()
+                await db.createLocation(testLocation1Id, '0', testLocation1PhoneNumber, MOV_THRESHOLD, 15, 0.2, 5000, 5000, 0, '+15005550006', '+15005550006', '+15005550006', 1000)
+                await door(testLocation1Id, ST_DOOR_STATUS.CLOSED)
+            })
+    
+            afterEach(async function() {
+                await redis.clearKeys()
+                await db.clearSessions()
+                await db.clearLocations()
+                helpers.log('\n')
+            })
+
+            it('should return 200 for a valid request', async ()=> {
+                const goodRequest = {
+                    locationid: testLocation1Id,
+                    signal: ST_DOOR_STATUS.CLOSED
+                }
+            
+                const response = await chai.request(server).post('/api/doorTest').send(goodRequest)
+                expect(response).to.have.status(200)
+            })
+
+            it('should return 400 for a request that does not contain locationid', async () => {
+                const badRequest = {
+                    signal: ST_DOOR_STATUS.CLOSED
+                }
+            
+                const response = await chai.request(server).post('/api/doorTest').send(badRequest)
+                expect(response).to.have.status(400)
+            })
+
+            it('should return 400 for a request that does not contain signal', async () => {
+                const badRequest = {
+                    locationid: testLocation1Id
+                }
+            
+                const response = await chai.request(server).post('/api/doorTest').send(badRequest)
+                expect(response).to.have.status(400)
+            })
+        })
+
+        describe('api/door endpoint', () => {
+            beforeEach(async function() {
+                await redis.clearKeys()
+                await db.clearSessions()
+                await db.clearLocations()
+                await db.createLocation(testLocation1Id, '0', testLocation1PhoneNumber, MOV_THRESHOLD, 15, 0.2, 5000, 5000, 0, '+15005550006', '+15005550006', '+15005550006', 1000, 'locationName', door_coreID, 'radar_particlecoreid')
+                await im21Door(door_coreID, IM21_DOOR_STATUS.CLOSED)
+            })
+        
+            afterEach(async function() {
+                await redis.clearKeys()
+                await db.clearSessions()
+                await db.clearLocations()
+                helpers.log('\n')
+            })
+            
+            it('should return 200 for a valid request', async ()=> {
+                const goodRequest = {
+                    coreid: door_coreID,
+                    data: `{ "deviceid": "FA:E6:51", "data": "${IM21_DOOR_STATUS.CLOSED}", "control": "86"}`
+                }
+            
+                const response = await chai.request(server).post('/api/door').send(goodRequest)
+                expect(response).to.have.status(200)
+            })
+
+            it('should return 400 for a request that does not contain coreid', async () => {
+                const badRequest = {
+                    data: `{ "deviceid": "FA:E6:51", "data": "${IM21_DOOR_STATUS.CLOSED}", "control": "86"}`
+                }
+            
+                const response = await chai.request(server).post('/api/door').send(badRequest)
+                expect(response).to.have.status(400)
+            })
+
+            it('should return 400 for a request that does not contain data', async () => {
+                const badRequest = {
+                    coreid: door_coreID
+                }
+            
+                const response = await chai.request(server).post('/api/door').send(badRequest)
+                expect(response).to.have.status(400)
+            })
+        })
+
+        describe('api/devicevitals endpoint', () => {
+            beforeEach(async function() {
+                await redis.clearKeys()
+                await db.clearSessions()
+                await db.clearLocations()
+                await db.createLocation(testLocation1Id, '0', testLocation1PhoneNumber, MOV_THRESHOLD, 15, 0.2, 5000, 5000, 0, '+15005550006', '+15005550006', '+15005550006', 1000, 'locationName', door_coreID, 'radar_particlecoreid')
+                await im21Door(door_coreID, IM21_DOOR_STATUS.CLOSED)
+            })
+        
+            afterEach(async function() {
+                await redis.clearKeys()
+                await db.clearSessions()
+                await db.clearLocations()
+                helpers.log('\n')
+            })
+
+
+            it('should return 200 for a valid request', async ()=> {
+                const goodRequest = {
+                    coreid: door_coreID,
+                    data: `{"device":{"network":{"signal":{"at":"Wi-Fi","strength":100,"strength_units":"%","strengthv":-47,"strengthv_units":"dBm","strengthv_type":"RSSI","quality":100,"quality_units":"%","qualityv":43,"qualityv_units":"dB","qualityv_type":"SNR"}},"cloud":{"connection":{"status":"connected","error":17,"attempts":1,"disconnects":9,"disconnect_reason":"error"},"coap":{"transmit":1305228,"retransmit":1721,"unack":0,"round_trip":1001},"publish":{"rate_limited":0}},"system":{"uptime":1298620,"memory":{"used":95000,"total":160488}}},"service":{"device":{"status":"ok"},"cloud":{"uptime":94305,"publish":{"sent":93201}},"coap":{"round_trip":1327}}}`
+                }
+            
+                const response = await chai.request(server).post('/api/devicevitals').send(goodRequest)
+                expect(response).to.have.status(200)
+            })
+
+            it('should return 400 for a request that does not contain coreid', async () => {
+                const badRequest = {
+                    data: `{"device":{"network":{"signal":{"at":"Wi-Fi","strength":100,"strength_units":"%","strengthv":-47,"strengthv_units":"dBm","strengthv_type":"RSSI","quality":100,"quality_units":"%","qualityv":43,"qualityv_units":"dB","qualityv_type":"SNR"}},"cloud":{"connection":{"status":"connected","error":17,"attempts":1,"disconnects":9,"disconnect_reason":"error"},"coap":{"transmit":1305228,"retransmit":1721,"unack":0,"round_trip":1001},"publish":{"rate_limited":0}},"system":{"uptime":1298620,"memory":{"used":95000,"total":160488}}},"service":{"device":{"status":"ok"},"cloud":{"uptime":94305,"publish":{"sent":93201}},"coap":{"round_trip":1327}}}`
+                }
+            
+                const response = await chai.request(server).post('/api/devicevitals').send(badRequest)
+                expect(response).to.have.status(400)
+            })
+
+            it('should return 400 for a request that does not contain data', async () => {
+                const badRequest = {
+                    coreid: door_coreID
+                }
+            
+                const response = await chai.request(server).post('/api/devicevitals').send(badRequest)
+                expect(response).to.have.status(400)
+            })
+
+            it('should return 400 for a request that contains a valid coreid and a totally invalid data field', async () => {
+                const badRequest = {
+                    coreid: door_coreID,
+                    data: `{"uselessField":"useless"}`
+                }
+            
+                const response = await chai.request(server).post('/api/devicevitals').send(badRequest)
+                expect(response).to.have.status(400)
+            })
+
+            it('should return 400 for a request that contains a valid coreid and an invalid data field missing signal strength', async () => {
+                const badRequest = {
+                    coreid: door_coreID,
+                    data: `{"device":{"network":{"signal":{"at":"Wi-Fi","strength_units":"%","strengthv":-47,"strengthv_units":"dBm","strengthv_type":"RSSI","quality":100,"quality_units":"%","qualityv":43,"qualityv_units":"dB","qualityv_type":"SNR"}},"cloud":{"connection":{"status":"connected","error":17,"attempts":1,"disconnects":9,"disconnect_reason":"error"},"coap":{"transmit":1305228,"retransmit":1721,"unack":0,"round_trip":1001},"publish":{"rate_limited":0}},"system":{"uptime":1298620,"memory":{"used":95000,"total":160488}}},"service":{"device":{"status":"ok"},"cloud":{"uptime":94305,"publish":{"sent":93201}},"coap":{"round_trip":1327}}}`
+                }
+            
+                const response = await chai.request(server).post('/api/devicevitals').send(badRequest)
+                expect(response).to.have.status(400)
+            })
+
+            it('should return 400 for a request that contains a valid coreid and an invalid data field missing device disconnects', async () => {
+                const badRequest = {
+                    coreid: door_coreID,
+                    data: `{{"device":{"network":{"signal":{"at":"Wi-Fi","strength":100,"strength_units":"%","strengthv":-47,"strengthv_units":"dBm","strengthv_type":"RSSI","quality":100,"quality_units":"%","qualityv":43,"qualityv_units":"dB","qualityv_type":"SNR"}},"cloud":{"connection":{"status":"connected","error":17,"attempts":1,"disconnect_reason":"error"},"coap":{"transmit":1305228,"retransmit":1721,"unack":0,"round_trip":1001},"publish":{"rate_limited":0}},"system":{"uptime":1298620,"memory":{"used":95000,"total":160488}}},"service":{"device":{"status":"ok"},"cloud":{"uptime":94305,"publish":{"sent":93201}},"coap":{"round_trip":1327}}}`
+                }
+            
+                const response = await chai.request(server).post('/api/devicevitals').send(badRequest)
+                expect(response).to.have.status(400)
+            })
+        })
+    })
+
 })